Resolve comment

+++ library/src/main/java/whiz/sspark/library/data/entity/Term.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Term(
         @SerializedName("id")val id: String = "",
-        @SerializedName("room") val room: Long? = null,
+        @SerializedName("room") val room: Int? = null,
         @SerializedName("term") val term: Long = 0,
         @SerializedName("year") val year: Long = 0,
         @SerializedName("academicGrade") val academicGrade: Int? = null
